+++ packages/server/api/src/app/database/migration/postgres/1751989232042-AddMcpsEnabled.ts
@@ -1,9 +1,14 @@
+import { ApEdition } from '@activepieces/shared'
 import { MigrationInterface, QueryRunner } from 'typeorm'
+import { isNotOneOfTheseEditions } from '../../database-common'
 
 export class AddMcpsEnabled1751989232042 implements MigrationInterface {
     name = 'AddMcpsEnabled1751989232042'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "platform_plan" DROP COLUMN "alertsEnabled"
         `)
@@ -23,6 +28,9 @@ export class AddMcpsEnabled1751989232042 implements MigrationInterface {
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "platform_plan" DROP COLUMN "mcpsEnabled"
         `)
